+++ deno/lib/__tests__/template-literal.test.ts
@@ -105,6 +105,7 @@ const stringStartsWithMax5 = z
 const brandedString = z
   .templateLiteral()
   .addInterpolatedPosition(z.string().min(1).brand("myBrand"));
+const anything = z.templateLiteral().addInterpolatedPosition(z.any());
 
 const url = z
   .templateLiteral()
@@ -201,6 +202,7 @@ test("template literal type inference", () => {
   util.assertEqual<z.infer<typeof stringMin5Max10>, string>(true);
   util.assertEqual<z.infer<typeof stringStartsWithMax5>, string>(true);
   util.assertEqual<z.infer<typeof brandedString>, string>(true);
+  util.assertEqual<z.infer<typeof anything>, `${any}`>(true);
 
   util.assertEqual<
     z.infer<typeof url>,
@@ -245,8 +247,6 @@ test("template literal unsupported args", () => {
       z.union([z.object({}), z.string()])
     )
   ).toThrow();
-  // @ts-expect-error
-  expect(() => z.templateLiteral().addInterpolatedPosition(z.any())).toThrow();
   expect(() =>
     // @ts-expect-error
     z.templateLiteral().addInterpolatedPosition(z.never())
@@ -328,6 +328,14 @@ test("template literal unsupported args", () => {
   expect(() =>
     // @ts-expect-error
     z.templateLiteral().addInterpolatedPosition(z.unknown())
+  ).toThrow();
+  expect(() =>
+    // @ts-expect-error
+    z.templateLiteral().addInterpolatedPosition(z.void())
+  ).toThrow();
+  expect(() =>
+    // @ts-expect-error
+    z.templateLiteral().addInterpolatedPosition(z.lazy(() => z.string()))
   ).toThrow();
   expect(() =>
     z.templateLiteral().addInterpolatedPosition(
@@ -424,6 +432,8 @@ test("template literal parsing - success - basic cases", () => {
   stringMin5Max10.parse("hello worl");
   stringStartsWithMax5.parse("hello");
   brandedString.parse("branded string");
+  anything.parse("");
+  anything.parse("everything");
 });
 
 test("template literal parsing - failure - basic cases", () => {
