+++ deno/lib/types.ts
@@ -29,6 +29,8 @@ import {
   ZodErrorMap,
   ZodIssue,
   ZodIssueCode,
+  ZodTemplateLiteralUnsupportedCheckError,
+  ZodTemplateLiteralUnsupportedTypeError,
 } from "./ZodError.ts";
 
 export { ZodParsedType } from "./helpers/util.ts";
@@ -5576,44 +5578,12 @@ export interface ZodTemplateLiteralDef extends ZodTypeDef {
   typeName: ZodFirstPartyTypeKind.ZodTemplateLiteral;
 }
 
-export class ZodTemplateLiteralUnsupportedTypeError extends Error {
-  constructor() {
-    super("Unsupported zod type!");
-
-    const actualProto = new.target.prototype;
-    if (Object.setPrototypeOf) {
-      // eslint-disable-next-line ban/ban
-      Object.setPrototypeOf(this, actualProto);
-    } else {
-      (this as any).__proto__ = actualProto;
-    }
-    this.name = "ZodTemplateLiteralUnsupportedTypeError";
-  }
-}
-
-export class ZodTemplateLiteralUnsupportedCheckError extends Error {
-  constructor(typeKind: ZodFirstPartyTypeKind, check: string) {
-    super(
-      `${typeKind}'s ${check} check is not supported in template literals!`
-    );
-
-    const actualProto = new.target.prototype;
-    if (Object.setPrototypeOf) {
-      // eslint-disable-next-line ban/ban
-      Object.setPrototypeOf(this, actualProto);
-    } else {
-      (this as any).__proto__ = actualProto;
-    }
-    this.name = "ZodTemplateLiteralUnsupportedCheckError";
-  }
-}
-
 export class ZodTemplateLiteral<Template extends string = ""> extends ZodType<
   Template,
   ZodTemplateLiteralDef
 > {
   addInterpolatedPosition<I extends TemplateLiteralInterpolatedPosition>(
-    type: Exclude<I, ZodNever | ZodAny | ZodNaN | ZodPipeline<any, any>>
+    type: Exclude<I, ZodNever | ZodNaN | ZodPipeline<any, any> | ZodLazy<any>>
   ): ZodTemplateLiteral<appendToTemplateLiteral<Template, I>> {
     return this._addPart(type) as any;
   }
